Support BO tuning in OpenMPOpt, fixes
- Fix CUGetMemInfo input type

+++ openmp/libomptarget/plugins-nextgen/cuda/src/rtl.cpp
@@ -566,7 +566,7 @@ struct CUDADeviceTy : public GenericDeviceTy {
     CUmemAllocationProp Prop = {};
     size_t Granularity = 0;
 
-    unsigned int Free, Total;
+    size_t Free, Total;
     CUresult Res = cuMemGetInfo(&Free, &Total);
     if (auto Err = Plugin::check(Res, "Error in cuMemGetInfo: %s"))
         return std::move(Err);
